efficientnet/requirements.txt
# conda install pytorch==1.8.0 torchvision torchaudio cudatoolkit=10.2
# conda install pytorch==1.8.0 torchvision==0.9.0 torchaudio==0.8.0 cpuonly
pyqt5
matplotlib
scikit-learn
timm
albumentations
pillow==8.4.0
opencv-python
pandas
tqdm
torchstat